fix(voting): arregla test votaciones
Arregla los ultimos fallos de los tests de points y multiple
